requirements.txt
# Core
numpy>=1.24.0
pandas>=2.0.0

# NLP + preprocessing
nltk>=3.8.1
sentence-transformers>=2.2.2
torch>=2.0.0

# Topic Modeling / Dimensionality Reduction
umap-learn>=0.5.3
hdbscan>=0.8.33

# Machine learning utils
scikit-learn>=1.3.0